// client/src/theme/types.ts
export type Colors = {
  white: string;
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  success: string;
  failure: string;
  warning: string;
  background: string;
  backgroundDisabled: string;
  backgroundAlt:string;
  text: string;
  textDisabled: string;
  textSubtle: string;
  borderColor: string;
  borderLight: string
  borderSecondary: string
  chatBgColor: string
};